Added new output for DeleteInstance operation. Small refactoring of constant classes.
