+++ cs-alibaba/src/main/java/io/cloudslang/content/alibaba/utils/constants/ExceptionMessages.java
@@ -1,0 +1,5 @@
+package io.cloudslang.content.alibaba.utils.constants;
+
+public class ExceptionMessages {
+    public static final String DELETE_INSTANCE_EXCEPTION = "Failed to delete instance.";
+}
