+++ cs-alibaba/src/main/java/io/cloudslang/content/alibaba/utils/constants/SuccessMessages.java
@@ -1,0 +1,5 @@
+package io.cloudslang.content.alibaba.utils.constants;
+
+public class SuccessMessages {
+    public static final String DELETE_INSTANCE_SUCCESS = "The instance delete request was successfully executed.";
+}
